Move HadoopDelegationTokenProvider
