+++ docs/security.md
@@ -758,7 +758,7 @@ explicitly provided to Spark at launch time. This is done by listing them in the
 
 Spark also supports custom delegation token providers using the Java Services
 mechanism (see `java.util.ServiceLoader`). Implementations of
-`org.apache.spark.deploy.security.HadoopDelegationTokenProvider` can be made available to Spark
+`org.apache.spark.security.HadoopDelegationTokenProvider` can be made available to Spark
 by listing their names in the corresponding file in the jar's `META-INF/services` directory.
 
 Delegation token support is currently only supported in YARN and Mesos modes. Consult the
